renamed committeesApi and removed unused prop in SendCommitteeMessage
