--- lib/api/committees.ts
@@ -1,3 +1,0 @@
-export const fetchOwCommittees = async () => {
-  return fetch(`/api/periods/ow-committees`).then((res) => res.json());
-};
